api for editing equipment status
