+++ internal/handlers/equipment_status.go
@@ -3,108 +3,72 @@ package handlers
 import (
 	"net/http"
 
-	"github.com/go-openapi/runtime/middleware"
-	"go.uber.org/zap"
-
 	"git.epam.com/epm-lstr/epm-lstr-lc/be/internal/generated/ent"
 	"git.epam.com/epm-lstr/epm-lstr-lc/be/internal/generated/swagger/models"
 	"git.epam.com/epm-lstr/epm-lstr-lc/be/internal/generated/swagger/restapi/operations"
-	eqStatusName "git.epam.com/epm-lstr/epm-lstr-lc/be/internal/generated/swagger/restapi/operations/equipment_status_name"
+	eqStatus "git.epam.com/epm-lstr/epm-lstr-lc/be/internal/generated/swagger/restapi/operations/equipment_status"
 	"git.epam.com/epm-lstr/epm-lstr-lc/be/internal/repositories"
 	"git.epam.com/epm-lstr/epm-lstr-lc/be/pkg/domain"
+	"github.com/go-openapi/runtime/middleware"
+	"github.com/go-openapi/strfmt"
+	"go.uber.org/zap"
 )
 
-func SetEquipmentStatusNameHandler(logger *zap.Logger, api *operations.BeAPI) {
-	equipmentStatusNameRepo := repositories.NewEquipmentStatusNameRepository()
-	statusHandler := NewEquipmentStatusName(logger)
+func SetEquipmentStatusHandler(logger *zap.Logger, api *operations.BeAPI) {
+	equipmentStatusRepo := repositories.NewEquipmentStatusRepository()
+	statusHandler := NewEquipmentStatus(logger)
 
-	api.EquipmentStatusNamePostEquipmentStatusNameHandler = statusHandler.PostEquipmentStatusNameFunc(equipmentStatusNameRepo)
-	api.EquipmentStatusNameListEquipmentStatusNamesHandler = statusHandler.ListEquipmentStatusNamesFunc(equipmentStatusNameRepo)
-	api.EquipmentStatusNameGetEquipmentStatusNameHandler = statusHandler.GetEquipmentStatusNameFunc(equipmentStatusNameRepo)
-	api.EquipmentStatusNameDeleteEquipmentStatusNameHandler = statusHandler.DeleteEquipmentStatusNameFunc(equipmentStatusNameRepo)
+	api.EquipmentStatusUpdateEquipmentStatusHandler = statusHandler.PutEquipmentStatusFunc(equipmentStatusRepo)
 }
 
-type EquipmentStatusName struct {
+type EquipmentStatus struct {
 	logger *zap.Logger
 }
 
-func NewEquipmentStatusName(logger *zap.Logger) *EquipmentStatusName {
-	return &EquipmentStatusName{
+func NewEquipmentStatus(logger *zap.Logger) *EquipmentStatus {
+	return &EquipmentStatus{
 		logger: logger,
 	}
 }
 
-func (c EquipmentStatusName) PostEquipmentStatusNameFunc(repository domain.EquipmentStatusNameRepository) eqStatusName.PostEquipmentStatusNameHandlerFunc {
-	return func(s eqStatusName.PostEquipmentStatusNameParams, access interface{}) middleware.Responder {
+func (c EquipmentStatus) PutEquipmentStatusFunc(eqStatusRepository domain.EquipmentStatusRepository) eqStatus.UpdateEquipmentStatusHandlerFunc {
+	return func(s eqStatus.UpdateEquipmentStatusParams, access interface{}) middleware.Responder {
 		ctx := s.HTTPRequest.Context()
-		name := s.Name.Name
-		createdStatus, err := repository.Create(ctx, *name)
+		statusName := s.Name.StatusName
+		data := models.EquipmentStatus{}
+		data.EndDate = s.Name.EndDate
+		data.StartDate = s.Name.StartDate
+		data.StatusName = statusName
+		data.ID = &s.ID
+
+		updatedEqStatus, err := eqStatusRepository.Update(ctx, &data)
 		if err != nil {
 			c.logger.Error("create status failed", zap.Error(err))
-			return eqStatusName.NewPostEquipmentStatusNameDefault(http.StatusInternalServerError).
+			return eqStatus.NewUpdateEquipmentStatusDefault(http.StatusInternalServerError).
 				WithPayload(buildStringPayload("can't create status"))
 		}
 
-		return eqStatusName.NewPostEquipmentStatusNameCreated().WithPayload(&models.SuccessEquipmentStatusNameOperationResponse{
-			Data: mapEquipmentStatusName(createdStatus),
+		id := int64(updatedEqStatus.ID)
+		endTime := strfmt.DateTime(updatedEqStatus.EndDate)
+		startTime := strfmt.DateTime(updatedEqStatus.StartDate)
+
+		return eqStatus.NewUpdateEquipmentStatusOK().WithPayload(&models.EquipmentStatusUpdateResponse{
+			ID:         &id,
+			EndDate:    &endTime,
+			StartDate:  &startTime,
+			StatusName: statusName,
 		})
 	}
 }
 
-func (c EquipmentStatusName) ListEquipmentStatusNamesFunc(repository domain.EquipmentStatusNameRepository) eqStatusName.ListEquipmentStatusNamesHandlerFunc {
-	return func(s eqStatusName.ListEquipmentStatusNamesParams, access interface{}) middleware.Responder {
-		ctx := s.HTTPRequest.Context()
-		statuses, err := repository.GetAll(ctx)
-		if err != nil {
-			c.logger.Error("get statuses failed", zap.Error(err))
-			return eqStatusName.NewListEquipmentStatusNamesDefault(http.StatusInternalServerError).
-				WithPayload(buildStringPayload("can't get statuses"))
-		}
-		listStatuses := models.ListEquipmentStatusNames{}
-		for _, element := range statuses {
-			listStatuses = append(listStatuses, mapEquipmentStatusName(element))
-		}
-		return eqStatusName.NewListEquipmentStatusNamesOK().WithPayload(listStatuses)
-	}
-}
-
-func (c EquipmentStatusName) GetEquipmentStatusNameFunc(repository domain.EquipmentStatusNameRepository) eqStatusName.GetEquipmentStatusNameHandlerFunc {
-	return func(s eqStatusName.GetEquipmentStatusNameParams, access interface{}) middleware.Responder {
-		ctx := s.HTTPRequest.Context()
-		foundStatus, err := repository.Get(ctx, int(s.StatusID))
-		if err != nil {
-			c.logger.Error("get status failed", zap.Error(err))
-			return eqStatusName.NewGetEquipmentStatusNameDefault(http.StatusInternalServerError).
-				WithPayload(buildStringPayload("can't get status"))
-		}
-
-		return eqStatusName.NewGetEquipmentStatusNameOK().WithPayload(&models.SuccessEquipmentStatusNameOperationResponse{
-			Data: mapEquipmentStatusName(foundStatus),
-		})
-	}
-}
-
-func (c EquipmentStatusName) DeleteEquipmentStatusNameFunc(repository domain.EquipmentStatusNameRepository) eqStatusName.DeleteEquipmentStatusNameHandlerFunc {
-	return func(s eqStatusName.DeleteEquipmentStatusNameParams, access interface{}) middleware.Responder {
-		ctx := s.HTTPRequest.Context()
-		deletedStatus, err := repository.Delete(ctx, int(s.StatusID))
-		if err != nil {
-			c.logger.Error("delete status failed", zap.Error(err))
-			return eqStatusName.NewDeleteEquipmentStatusNameDefault(http.StatusInternalServerError).
-				WithPayload(buildStringPayload("can't delete status"))
-		}
-		return eqStatusName.NewDeleteEquipmentStatusNameOK().WithPayload(
-			&models.SuccessEquipmentStatusNameOperationResponse{
-				Data: mapEquipmentStatusName(deletedStatus),
-			},
-		)
-	}
-}
-
-func mapEquipmentStatusName(status *ent.EquipmentStatusName) *models.EquipmentStatusName {
-	id := int64(status.ID)
-	return &models.EquipmentStatusName{
-		ID:   id,
-		Name: &status.Name,
+func mapEquipmentStatus(eqStatus *ent.EquipmentStatus) *models.EquipmentStatusUpdateResponse {
+	id := int64(eqStatus.ID)
+	endTime := strfmt.DateTime(eqStatus.EndDate)
+	startTime := strfmt.DateTime(eqStatus.StartDate)
+	return &models.EquipmentStatusUpdateResponse{
+		ID:         &id,
+		EndDate:    &endTime,
+		StartDate:  &startTime,
+		StatusName: &eqStatus.Edges.EquipmentStatusName.Name,
 	}
 }
